+++ cpp/include/copc-lib/io/copc_writer.hpp
@@ -7,7 +7,7 @@
 #include <stdexcept>
 #include <string>
 
-#include "copc-lib/copc/config.hpp"
+#include "copc-lib/copc/copc_config.hpp"
 #include "copc-lib/geometry/box.hpp"
 #include "copc-lib/io/copc_base_io.hpp"
 #include "copc-lib/las/header.hpp"