+++ cpp/include/copc-lib/io/internal/copc_writer_internal.hpp
@@ -3,7 +3,7 @@
 
 #include <ostream>
 
-#include "copc-lib/copc/config.hpp"
+#include "copc-lib/copc/copc_config.hpp"
 #include "copc-lib/io/copc_base_io.hpp"
 #include "copc-lib/las/header.hpp"