+++ cpp/src/io/laz_base_writer.cpp
@@ -1,0 +1,1 @@
+#include "copc-lib/io/laz_base_writer.hpp"